Rework index.js to be more readable.
This is subjective, but I've found that using two classes (one private, one
public) and a method to tie them to each other, was much more readable than
passing instances to functions.

This adds a new private class, tied to the public instance, that is deffered to
for all public functions. The public functions return either undefined, or the
public instance for the result of the private call.